Move Enforcement::Notification into the EnforcementPresenter
