+++ app/presenters/enforcement_presenter.rb
@@ -31,6 +31,13 @@ class EnforcementPresenter
     end
   end
 
+  def start_investigation_email
+    {
+      subject: start_investigation_email_subject,
+      body: start_investigation_email_body
+    }
+  end
+
   def close_investigation_email(**)
     {
       subject: close_investigation_email_subject,
@@ -39,6 +46,26 @@ class EnforcementPresenter
   end
 
   private
+
+  def officer_email
+    case_record.user_email || local_authority.feedback_email || local_authority.email_address
+  end
+
+  def start_investigation_email_subject
+    I18n.t("bops_enforcements.start_investigation_email.subject", ref: case_record.id)
+  end
+
+  def start_investigation_email_body
+    I18n.t("bops_enforcements.start_investigation_email.body",
+      ref: case_record.id,
+      address: address.to_s,
+      received_on: I18n.l(received_at.to_date),
+      report_date: I18n.l(received_at.to_date),
+      complainant_name: complainant.name,
+      days: 20,
+      officer_email:,
+      council_name: local_authority.council_name)
+  end
 
   def close_investigation_email_subject
     I18n.t("bops_enforcements.close_investigation_email.subject", ref: case_record.id)
@@ -64,7 +91,7 @@ class EnforcementPresenter
       complainant_name: complainant.name,
       reason:,
       additional_comment:,
-      officer_email: case_record.user_email,
+      officer_email:,
       council_name: local_authority.council_name)
   end
 end
